refactor: split RFID scanning into domain processor and UI presenter
