+++ src/hooks/useRfidScanning.ts
@@ -1,275 +1,35 @@
 import { adapter } from '@platform';
 import { useEffect, useRef, useCallback } from 'react';
 
+import {
+  isMockScanSourceRunning,
+  resetMockScanSourceForTesting,
+  startMockScanSource,
+  stopMockScanSource,
+} from '../dev/mockScanSource';
 import { isRealScanningEnabled } from '../platform/adapter';
 import type { NfcScanEvent } from '../platform/adapter';
-import { api, mapApiErrorToGerman, ApiError, formatRoomName } from '../services/api';
-import type { RfidScanResult, CurrentSession } from '../services/api';
+import { api } from '../services/api';
+import type { RfidScanResult } from '../services/api';
+import {
+  createScanErrorResult,
+  createSessionExpiredResult,
+  createSupervisorRedirectResult,
+  evaluateSupervisorScan,
+  processStudentBookkeeping,
+  runPickupQuery,
+  updateSessionActivityQuietly,
+} from '../services/scanProcessor';
 import { useUserStore } from '../store/userStore';
-import { getSecureRandomInt } from '../utils/crypto';
 import { createLogger, serializeError } from '../utils/logger';
-
-// Mock scanning interval for development
-let mockScanInterval: ReturnType<typeof setInterval> | null = null;
-let mockScanCounter = 0;
 
 /** Reset module-level state between tests. Not for production use. */
 export function __resetModuleStateForTesting(): void {
-  if (mockScanInterval) {
-    clearInterval(mockScanInterval);
-    mockScanInterval = null;
-  }
-  mockScanCounter = 0;
+  resetMockScanSourceForTesting();
 }
 
 const logger = createLogger('useRfidScanning');
 const PROCESSED_SCAN_ID_TTL_MS = 30_000;
-
-// ============================================================================
-// Helper functions to reduce cognitive complexity in processScan
-// ============================================================================
-
-/**
- * Creates a session expired error result for display.
- */
-const createSessionExpiredResult = (): RfidScanResult & { navigateOnClose: string } => ({
-  student_name: 'Sitzung abgelaufen',
-  student_id: null,
-  action: 'error',
-  message: 'Bitte melden Sie sich erneut an.',
-  showAsError: true,
-  navigateOnClose: '/home',
-});
-
-/**
- * Handles supervisor authentication from RFID scan.
- * Returns true if supervisor was handled (caller should return early).
- */
-interface SupervisorHandlerParams {
-  result: RfidScanResult;
-  tagId: string;
-  currentSession: CurrentSession | null;
-  pin: string;
-  scannedSupervisorsRef: { current: Set<number> };
-  addSupervisorFromRfid: (staffId: number, staffName: string) => boolean;
-  addActiveSupervisorTag: (tagId: string) => void;
-  isActiveSupervisor: (tagId: string) => boolean;
-  setScanResult: (result: RfidScanResult) => void;
-  showScanModal: () => void;
-  showSupervisorRedirect: () => void;
-}
-
-const handleSupervisorAuthentication = async (
-  params: SupervisorHandlerParams
-): Promise<boolean> => {
-  const {
-    result,
-    tagId,
-    currentSession,
-    pin,
-    scannedSupervisorsRef,
-    addSupervisorFromRfid,
-    addActiveSupervisorTag,
-    isActiveSupervisor,
-    setScanResult,
-    showScanModal,
-    showSupervisorRedirect,
-  } = params;
-
-  if (result.action !== 'supervisor_authenticated') {
-    return false;
-  }
-
-  const staffId = result.student_id;
-  const staffName = result.student_name;
-
-  if (staffId === null) {
-    return false;
-  }
-
-  const alreadySelected = addSupervisorFromRfid(staffId, staffName);
-  const wasSeenThisSession = scannedSupervisorsRef.current.has(staffId);
-  const isRepeatSupervisor = alreadySelected || wasSeenThisSession || isActiveSupervisor(tagId);
-
-  // Track in-memory and mark tag active for fast return
-  scannedSupervisorsRef.current.add(staffId);
-  addActiveSupervisorTag(tagId);
-
-  // Sync supervisors with backend (fire-and-forget)
-  if (currentSession && pin) {
-    void syncSupervisorsWithBackend(currentSession, pin, staffId);
-  }
-
-  logger.info('Supervisor authenticated successfully', {
-    supervisorName: staffName,
-    message: result.message,
-    staffId,
-    isRepeatSupervisor,
-    alreadySelected,
-  });
-
-  // Show redirect for repeat supervisors
-  if (isRepeatSupervisor) {
-    showSupervisorRedirect();
-    return true;
-  }
-
-  // First-time supervisor scan - show added message
-  const firstScanResult: RfidScanResult = {
-    ...result,
-    student_name: 'Betreuer erkannt',
-    message: `${result.student_name} wurde als Betreuer zu diesem Raum hinzugefügt.`,
-  };
-  setScanResult(firstScanResult);
-  showScanModal();
-  return true;
-};
-
-/**
- * Syncs supervisor list with backend after RFID authentication.
- */
-const syncSupervisorsWithBackend = async (
-  currentSession: CurrentSession,
-  pin: string,
-  staffId: number
-): Promise<void> => {
-  try {
-    const updatedSupervisorIds = useUserStore.getState().selectedSupervisors.map(s => s.id);
-    await api.updateSessionSupervisors(pin, currentSession.active_group_id, updatedSupervisorIds);
-    logger.info('Supervisor synced via RFID (network path)', {
-      staffId,
-      sessionId: currentSession.active_group_id,
-    });
-  } catch (error) {
-    logger.warn('Supervisor sync failed (network path)', {
-      error: error instanceof Error ? error.message : String(error),
-      staffId,
-    });
-  }
-};
-
-/**
- * Processes successful student scan - updates the short-lived result cache.
- */
-interface StudentBookkeepingParams {
-  result: RfidScanResult;
-  tagId: string;
-  recordTagScan: (
-    tagId: string,
-    data: { timestamp: number; studentId?: string; result?: RfidScanResult }
-  ) => void;
-}
-
-const processStudentBookkeeping = (params: StudentBookkeepingParams): void => {
-  const { result, tagId, recordTagScan } = params;
-
-  if (!result.student_id) {
-    return;
-  }
-
-  const studentId = result.student_id.toString();
-
-  // Short-lived result cache (in-memory only)
-  recordTagScan(tagId, {
-    timestamp: Date.now(),
-    studentId,
-    result,
-  });
-};
-
-/**
- * Determines error title based on error type.
- */
-const getErrorTitle = (error: unknown): string => {
-  if (error instanceof ApiError) {
-    if (error.code === 'ROOM_CAPACITY_EXCEEDED') return 'Raum voll';
-    if (error.code === 'ACTIVITY_CAPACITY_EXCEEDED') return 'Aktivität voll';
-  }
-  return 'Scan fehlgeschlagen';
-};
-
-/**
- * Build the duplicate-active-visit modal copy. The backend (Issue #844)
- * includes `room_name` in details so we can tell the user the actual room
- * the student is already in — which is often NOT the room being scanned.
- *
- * Two important details:
- *   1. We pass `room_name` through `formatRoomName` so internal "WC"
- *      surfaces as "Toilette" (consistent with `mapApiErrorToGerman` and
- *      the rest of the kiosk UI).
- *   2. If `room_name` is missing (degraded 409 path: backend couldn't
- *      reload the existing visit, or it was just closed by another
- *      scan), we do NOT know that the student is in a different room —
- *      the lookup may simply have failed. Use the same neutral copy as
- *      `mapApiErrorToGerman`'s fallback rather than claiming "anderer
- *      Raum", which could send staff to the wrong place.
- */
-const buildAlreadyInMessage = (error: unknown): string => {
-  if (error instanceof ApiError) {
-    const roomName = error.details?.room_name;
-    if (typeof roomName === 'string' && roomName.length > 0) {
-      return `Bereits angemeldet in ${formatRoomName(roomName)}.`;
-    }
-  }
-  return 'Schüler*in ist bereits angemeldet.';
-};
-
-/**
- * Detect duplicate-active-visit responses. Prefer the structured
- * `STUDENT_ALREADY_ACTIVE` code from the new 409 body (Issue #844). Keep
- * the substring fallback so older backend builds without the structured
- * response still resolve to the friendly modal instead of a generic error.
- */
-const isStudentAlreadyActiveError = (error: unknown, errorMessage: string): boolean => {
-  if (error instanceof ApiError && error.code === 'STUDENT_ALREADY_ACTIVE') {
-    return true;
-  }
-  return errorMessage.includes('already has an active visit');
-};
-
-/**
- * Extract the student_id from a STUDENT_ALREADY_ACTIVE error if present.
- * The backend (Issue #844) ships the student id in `details.student_id`
- * even on the degraded 409 path, so we always have at least the identity
- * to wire into the dedup map. Older backends that only emit the substring
- * fallback won't include details — return null and accept that the next
- * scan event will hit the backend again until a successful scan
- * populates the mapping organically.
- */
-const extractAlreadyActiveStudentId = (error: unknown): number | null => {
-  if (error instanceof ApiError && typeof error.details?.student_id === 'number') {
-    return error.details.student_id;
-  }
-  return null;
-};
-
-/**
- * Creates an error result for display when scan fails.
- */
-const createScanErrorResult = (error: unknown): RfidScanResult => {
-  const errorMessage = error instanceof Error ? error.message : String(error);
-
-  // Special handling for "already checked in" scenario
-  if (isStudentAlreadyActiveError(error, errorMessage)) {
-    return {
-      student_name: 'Bereits eingecheckt',
-      student_id: extractAlreadyActiveStudentId(error),
-      action: 'already_in',
-      message: buildAlreadyInMessage(error),
-      isInfo: true,
-    };
-  }
-
-  // Generic error handling
-  const userFriendlyMessage = mapApiErrorToGerman(error);
-  return {
-    student_name: getErrorTitle(error),
-    student_id: null,
-    action: 'error',
-    message: userFriendlyMessage || 'Bitte erneut versuchen',
-    showAsError: true,
-  };
-};
 
 export const useRfidScanning = () => {
   // Note: Navigation is now handled by page components via navigateOnClose flag in scan result
@@ -300,16 +60,7 @@ export const useRfidScanning = () => {
   const processedScanIdsRef = useRef<Map<number, number>>(new Map());
 
   const showSupervisorRedirect = useCallback(() => {
-    const redirectResult: RfidScanResult = {
-      student_name: 'Betreuer erkannt',
-      student_id: null,
-      action: 'supervisor_authenticated',
-      message: 'Betreuer wird zum Home-Bildschirm weitergeleitet.',
-      isInfo: true,
-      // Flag to indicate navigation should happen on modal close
-      navigateOnClose: '/home',
-    } as RfidScanResult & { navigateOnClose: string };
-    setScanResult(redirectResult);
+    setScanResult(createSupervisorRedirectResult());
     showScanModal();
     // Modal timeout and navigation handled by ActivityScanningPage via useModalTimeout
   }, [setScanResult, showScanModal]);
@@ -395,51 +146,24 @@ export const useRfidScanning = () => {
         lockPickupQueryTag(tagId);
         addToProcessingQueue(tagId);
         try {
-          const result = await api.queryPickupInfo({ student_rfid: tagId }, freshUser.pin);
-          const latestState = useUserStore.getState();
-          if (
-            latestState.rfid.scanMode !== 'pickupQuery' ||
-            latestState.rfid.scanContextId !== freshRfid.scanContextId
-          ) {
-            logger.debug('Discarding stale pickup query result', {
-              tagId,
-              scanContextId: freshRfid.scanContextId,
-            });
+          const outcome = await runPickupQuery({
+            tagId,
+            pin: freshUser.pin,
+            scanContextId: freshRfid.scanContextId,
+          });
+
+          if (outcome.status === 'stale') {
             return;
           }
 
-          const { active_students: _, ...pickupResult } = result;
-          setScanResult({ ...pickupResult, scannedTagId: tagId });
+          setScanResult(outcome.result);
           showScanModal();
-          removeFromProcessingQueue(tagId);
 
-          try {
-            await api.updateSessionActivity(freshUser.pin);
-            logger.debug('Session activity updated after pickup query');
-          } catch (activityError) {
-            logger.warn('Failed to update session activity after pickup query', {
-              error: activityError,
-            });
+          if (outcome.status === 'success') {
+            removeFromProcessingQueue(tagId);
+            await updateSessionActivityQuietly(freshUser.pin, 'pickup');
           }
 
-          return;
-        } catch (error) {
-          logger.error('Failed to query pickup info', { error: serializeError(error) });
-
-          const latestState = useUserStore.getState();
-          if (
-            latestState.rfid.scanMode !== 'pickupQuery' ||
-            latestState.rfid.scanContextId !== freshRfid.scanContextId
-          ) {
-            logger.debug('Discarding stale pickup query error', {
-              tagId,
-              scanContextId: freshRfid.scanContextId,
-            });
-            return;
-          }
-
-          setScanResult(createScanErrorResult(error));
-          showScanModal();
           return;
         } finally {
           removeFromProcessingQueue(tagId);
@@ -488,22 +212,25 @@ export const useRfidScanning = () => {
         // instead of looking it up from recentTagScans (fixes race condition)
         setScanResult({ ...result, scannedTagId: tagId });
 
-        // Handle supervisor authentication (returns true if handled)
-        const supervisorHandled = await handleSupervisorAuthentication({
+        // Handle supervisor authentication (handled outcomes end the scan here)
+        const supervisorOutcome = await evaluateSupervisorScan({
           result,
           tagId,
           currentSession: freshSession,
           pin: freshUser.pin,
-          scannedSupervisorsRef,
+          scannedSupervisors: scannedSupervisorsRef.current,
           addSupervisorFromRfid,
           addActiveSupervisorTag,
           isActiveSupervisor,
-          setScanResult,
-          showScanModal,
-          showSupervisorRedirect,
         });
 
-        if (supervisorHandled) {
+        if (supervisorOutcome.handled) {
+          if (supervisorOutcome.presentation === 'redirect') {
+            showSupervisorRedirect();
+          } else {
+            setScanResult(supervisorOutcome.result);
+            showScanModal();
+          }
           return;
         }
 
@@ -518,12 +245,7 @@ export const useRfidScanning = () => {
         showScanModal();
 
         // Update session activity (fire-and-forget)
-        try {
-          await api.updateSessionActivity(freshUser.pin);
-          logger.debug('Session activity updated');
-        } catch (activityError) {
-          logger.warn('Failed to update session activity', { error: activityError });
-        }
+        await updateSessionActivityQuietly(freshUser.pin, 'scan');
       } catch (error) {
         logger.error('Failed to process RFID scan', { error: serializeError(error) });
         const errorResult = createScanErrorResult(error);
@@ -647,41 +369,9 @@ export const useRfidScanning = () => {
       logger.info('RFID not enabled, starting mock scanning mode');
 
       // Start mock scanning interval
-      if (!mockScanInterval) {
+      if (!isMockScanSourceRunning()) {
         startRfidScanning(); // Update store state
-
-        // Generate mock scans every 3-5 seconds
-        mockScanInterval = setInterval(
-          () => {
-            // Get mock tags from environment variable or use defaults
-            const envTags = import.meta.env.VITE_MOCK_RFID_TAGS as string | undefined;
-            const mockStudentTags: string[] = envTags
-              ? envTags.split(',').map(tag => tag.trim())
-              : [
-                  // Default realistic hardware format tags
-                  '04:D6:94:82:97:6A:80',
-                  '04:A7:B3:C2:D1:E0:F5',
-                  '04:12:34:56:78:9A:BC',
-                  '04:FE:DC:BA:98:76:54',
-                  '04:11:22:33:44:55:66',
-                ];
-
-            // Pick a random tag from the list using unbiased secure randomness
-            const randomIndex = getSecureRandomInt(mockStudentTags.length);
-            const mockTagId = mockStudentTags[randomIndex];
-            const scanId = ++mockScanCounter;
-
-            logger.info('Mock RFID scan generated', {
-              tagId: mockTagId,
-              scanId,
-              platform: 'Development Mock',
-            });
-
-            onAdapterScan({ tagId: mockTagId, scanId });
-          },
-          5000 + getSecureRandomInt(5000)
-        ); // Random interval between 5-10 seconds
-
+        startMockScanSource(onAdapterScan);
         isServiceStartedRef.current = true;
         logger.info('Mock RFID scanning started');
       }
@@ -743,10 +433,9 @@ export const useRfidScanning = () => {
         if (!backendStopped) {
           logger.warn('RFID backend did not confirm stop within timeout');
         }
-      } else if (mockScanInterval) {
+      } else if (isMockScanSourceRunning()) {
         // Stop mock scanning
-        clearInterval(mockScanInterval);
-        mockScanInterval = null;
+        stopMockScanSource();
         logger.info('Mock RFID scanning stopped');
       }
       isServiceStartedRef.current = false;
@@ -815,10 +504,7 @@ export const useRfidScanning = () => {
   // Cleanup on unmount
   useEffect(() => {
     return () => {
-      if (mockScanInterval) {
-        clearInterval(mockScanInterval);
-        mockScanInterval = null;
-      }
+      stopMockScanSource();
       if (isServiceStartedRef.current) {
         void stopScanning();
       }
